refactor: update tab and taborder w/zustand store
